Episode02: check spelling in US English
Being the developers from the Netherlands eScience CentER, I have applied a
round of spellchecking in US English. This affected the body of the text as
well as the name of variables (labelled > labeled). This very broad type of
editing can be applied ahead of distinguishing linguistic and pedagogic
issues.