E02-language: revise up to 'Convolution on the GPU Using CuPy'
This revision has aimed at improving the learninge experience by means of:
(1) preserving content, barring a few sentences repeated within a short
range, and some incidental observations that could still be inferred from
the context;
(2) merging, splitting and relocating sentences in order to provide a
reading experience in consistent blocks, mostly ordered from general to
specific notions, so that a topic is opened and closed within a narrow
span of text;
(3) changing the text at word level to reduce repetitions and/or convey
more precision;
(4) only comments have been edited in the code lines; editing the code
lines is to be handled in a separate branch named E02-code.

Although conciseness and brevity are different notions, the word count went
from 4660 to 4568.

INFORMATION CHANGES

The revision deserves revision as far as meanings are concerned. Broadly
speaking, editing was supposed to streamline the reading/learning
experience and not to alter the content, which may have happened
unintentionally.

However, the following changes are deliberate and should be revised all the
more:
(1) the connection between CPU and GPU. Rather than with a cable. I propose
they are more likely connected with some interconnect plug. I have add a
Wikipedia link to interconnects and refrained to expand on hardware
specifications too widely, for example mentioning NVLink and developments.
(2) the status of cupyx as a namespace of CuPy. I even got somewhere that
once upon a time cupyx used to be referred to as cupy.cupyx. At any rate,
cupyx is documented inside Cupy. I have added a link to the official
documentation. My new wording wanted to convey the relatedness of CuPy and
cupyx and can be inaccurate, but also read (1) next.

FUTURE

Remarks not implemented but worth considering in the future are:
(1) clarifying from the outset overlaps and distinctions between the terms
API v libraries v packages v modules, and between methods v functions v
routines. Then, stick to a preferred usage nor to confuse learners neither
to reinforce confusions;
(2) if not done in a previous episode, mentioning that we are only coding
on Nvidia graphics cards because of the CUDA lock;
(3) revising this episode's questions and objectives (at the top of the
page) in a separate branch to be named E02-pedagogy.

ONE-OFF NOTE ON LINGUISTIC EDITING

As seen here, linguistic revision takes time and space. In terms of
psychology of copy editing, it would be unsurprising if the authors of the
original content feel uneasy about changes at first. I hope nobody takes
umbrage at that, although this is a ritual moment of copy editing. Exposing
the reasoning behind the changes is a handle to reframe the owner's view of
a text and evaluate the new state of the play with a dispassionate eye. The
risk of cutting corners is, naturally, to lose edge.