E02-language: revise the first episode half 'Convolution in Python'
The revision work of commit 9640c8c is extended to the section 'NumPy
routines on the GPU' included. The spirit is the same as in commit 9640c8c

Some edits leaked in the rest of the documents, either by programmatic
substitution or because of establishing an order of appearance
(abbreviations).

Left to do is the second half of the episode with the astronomy example.