E02-code: rename variable 'deltas' as 'diracs'
It seems clear that the previous variable name 'deltas' did not express
generic finite differences but an application of Dirac delta function to
the image pixels. 'diracs' is therefore more precise insofar as it rules
out any ambiguity. Since the convolution filter is called gauss, the pair
diracs/gauss sounds also well balanced. Any other choice is of course
possible, but without the drawbacks of the delta terminology