EO2-language/pedagogy: incomplete revision of the second episode half on Astronomy
Here linguistic editing has been largely functional to pedagogy improvements.

UNTIL THE SECTION ON BACKGROUND CHARACTERIZATION (STEP 1)

The style goes too conversational about specialist knowledge, and seems to
imply that just speaking of specialist knowledge is sufficient for learners
to focus on the applications in GPU programming. In fact, the current
implementation of a good idea approach engages uninitiated learners with
considerable extraneous load at the expense of the cognitive load. This may
prove quite tiresome during the course and while revising the material.

The editing work so far to limit the side effects of a specialist
application involved among others: (1) re-locate sentences so as to guide
the readers from more- to less-known notions; (2) add a few Wikipedia links
for exotic notions; (3) highlight/disclose assumptions, purposes and
conclusions for the uninitiated readers. Hopefully, the uninitiated reader
is better prepared to swallow the purpose of this section. Please double
check for correctness.

FROM SECTION ON SEGMENTING (STEP 2) ONWARD

The text and logic become too hard to unravel and follow, both
linguistically and pedagogically. The germane and extraneous loads for
learners are considerable and possibly excessive. Intervening on these
sections requires too much guesswork. Only a light linguistic pass has then
been applied as a foothold for suggested work.

OVERALL ON THE ASTRONOMY PART

After having read the text and code several times, I guess that the purpose
of this standard routine in radio astronomy is to find out sources lost
WITHIN the background, which is then translated into finding the sources
WITHIN the clipped dataset.

Realising this (or anything that applies instead) should not be guesswork
in the first place. The course should provide the necessary, accompanying
and lightweight information early on.

Also, the assisting commentary should be tuned in well with the overall
statement of purpose of the method. Unglued specialist claims (for example,
those on the probability of finding bright pixels at random) should fall
into place without raising questions (how did you compute that probability?
why is it equally relevant to the initial image of 2048² pixels and to the
clipped image, which is arguably smaller?)

The course should play ahead of these doubts. Shifting attention between
inconsistent/incomplete specialist claims muddles the notions of GPU
programming.

As a loadstar, the learner should recognize in the text -- as easily as
possible -- what to do, how to do that, and why to do that.

(This extended commit message may well become the explanation of a pull
request/issue.)