E02-language/code: typesetting convention functions/methods/variable/attributes
Using this native error message as example:

	Please use `.get()` to construct a NumPy array explicitly.

I have applied the following usage:

(1) Functions are typeset as `function()`
(2) Methods as either `.method()` or `module.method()`
(3) Variables as `variable`
(4) Attributes as `.attribute`

No decision made yet as to the Jupyter magics.

This is an example of the implementation

	The previous code executes `convolve2d_gpu()` ten times, and stores the
	execution time of each run, in seconds, inside the `.gpu_times` attribute
	of the variable `benchmark_gpu`. We can then compute the average execution
	time and print it, as shown.

For me: this commit shows the limit of the division into the branch
E02-code and E0-language. The edits in this commit are in-line code text,
so editing this code directly in the language-dedicated branch saves me one
merge with conflicts